feat(): created custom hook useRestaurant
