+++ FoodVilla/src/utils/helper.js
@@ -1,0 +1,5 @@
+export function filterData(searchText, restaurants) {
+  return restaurants.filter((restaurant) =>
+    restaurant?.data?.name?.toLowerCase().includes(searchText.toLowerCase())
+  );
+}
